+++ cogs/starboard.py
@@ -161,7 +161,7 @@ class Starboard(commands.Cog):
         await original_message.edit(embed=new_embed)
 
     @cog_ext.cog_subcommand(base="starboard", name="setup",
-                            description="Sets up a new starboard for this server in a given channel - updates a starboard if there already is one in the channel.",
+                            description="Sets up a starboard for a given channel - updates it if there already is one in the channel.",
                             options=[{"name": "channel", "description": "The channel to set the class up in.", "type": 7, "required": True},
                                      {"name": "emoji", "description": "The emoji for the starboard to use.", "type": 3, "required": True},
                                      {"name": "threshold", "description": "If a message has under these many stars, it will be automatically deleted.", "type": 4}],